Updated "9. Palindrome Number" (c++, java)

+++ 9. Palindrome Number/Java.java
@@ -2,7 +2,7 @@ public class Java {
 
     public static boolean isPalindrome(int x) {
         int n = x;
-        int revX = 0;
+        long revX = 0;
         while(n>0) {
             revX = revX*10 + n%10;
             n = n/10;
